feat: my portfolio revenue notification (#23)

+++ src/main/java/Back/whats_your_ETF/dto/PortfolioResponse.java
@@ -5,6 +5,6 @@ import lombok.*;
 public record PortfolioResponse(
         Long portfolioId,
         String title,
-        Long revenue,
+        Double revenue,
         Long investAmount
 ) {}
